+++ components/Elements/index.ts
@@ -1,5 +1,5 @@
 export { default as Banner } from './Banner';
-export { default as Button } from './Buttons/Button';
+export { default as Button } from './buttons/Button';
 export { default as CheckboxElement } from './CheckboxElement';
 export { default as Input } from './Input';
 export { default as Loading } from './Loading';
