+++ components/Modules/Map/functions/GetPoints.ts
