+++ components/Modules/Map/index.ts
@@ -1,0 +1,7 @@
+export { default as Clusters } from './Clusters';
+export { default as FetchMore } from './functions/FetchMore';
+export { default as GetPoints } from './functions/GetPoints';
+export { default as Map } from './Map';
+export { default as MapStyleMenu } from './MapStyleMenu';
+export { default as Marker } from './Marker';
+export { default as SearchLocation } from './SearchLocation';
