--- components/Modules/SearchLocation/index.ts
@@ -1,1 +1,0 @@
-export { default as SearchLocation } from './SearchLocation';
