--- components/Templates/MainPage/index.ts
@@ -1,1 +1,0 @@
-export { default as Search } from './search';
