+++ lib/functions/index.ts
@@ -2,5 +2,4 @@ export { default as FirstLetterUppercase } from './FirstLetterUppercase';
 export { default as GetCookieFromServerSideProps } from './GetCookieFromServerSideProps';
 export { default as IsJwtValid } from './IsJwtValid';
 export { default as LogOut } from './LogOut';
-export { default as ConvertBounds } from './map/ConvertBounds';
 export { default as SSRRedirect } from './SSRRedirect';
