+++ pages/user/[username]/collections.tsx
@@ -1,7 +1,6 @@
 import { ApolloClient, createHttpLink, InMemoryCache } from '@apollo/client';
-import { CollectionsIcon, PlusIcon } from '@components/icons';
 import UserLayout from '@components/Layouts/User';
-import Card from '@components/Modules/UserPage/Card';
+import Card from '@components/modules/UserPage/Card';
 import { UserDocument, UserQuery } from '@graphql/user.graphql';
 import {
   GetCookieFromServerSideProps,
@@ -12,6 +11,7 @@ import { GetServerSidePropsContext } from 'next';
 import Link from 'next/link';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
+import { HiFolderOpen, HiPlus } from 'react-icons/hi';
 
 import { ValidateUsername } from '../../../shared/validation';
 
@@ -36,13 +36,13 @@ const CollectionsPage: React.FC<{
       <div className="flex justify-end w-full">
         <Link href="/create/post" passHref>
           <button className="flex items-center px-3 py-1 text-xs font-semibold text-gray-500 border border-gray-500 rounded-full gap-2 space-x-1.5 shadown-sm w-max">
-            <PlusIcon className="w-2 h-2" /> {t('create_collection')}
+            <HiPlus className="w-2 h-2" /> {t('create_collection')}
           </button>
         </Link>
       </div>
       {userQuery.user.collections?.length == 0 ? (
         <Card>
-          <CollectionsIcon className="w-8 h-8" />
+          <HiFolderOpen className="w-8 h-8" />
           <div>{t('no_collections')}</div>
         </Card>
       ) : (
